+++ app/src/main/java/cityfreqs/com/pilfershushjammer/jammers/ActiveJammerService.java
@@ -101,7 +101,6 @@ public class ActiveJammerService extends Service {
         notifyActiveBuilder = new NotificationCompat.Builder(this, CHANNEL_ID);
 
         notifyActiveBuilder.setSmallIcon(R.mipmap.ic_stat_logo_notify_jammer)
-                //.setLargeIcon(BitmapFactory.decodeResource(this.getResources(), R.mipmap.ic_launcher))
                 .setContentTitle(getResources().getString(R.string.app_status_11))
                 .setContentText(getResources().getString(R.string.app_status_12))
                 .setContentIntent(pendingIntent)