+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * audioManager property support near ultrasound check
 * packageInfo exception handling
 * removed notification xml icon crashing some devices
+* build tools update
 
 ### 4.5.2 (2020-10-16)
 * bugfix for adding already added fragment